[fix] Headerをレスポンシブ対応
