+++ view/next-project/src/components/common/Header/Header.type.ts
@@ -1,0 +1,3 @@
+export interface HeaderProps {
+    onSideNavOpen?: () => void;
+}
